+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTV1B07KDQT6ZFYY2X11KGW9"
+        "01KTVC27ZHPZHTMX0A3FV9T0VJ"
     )
     r"""The unique identifier of the data source"""