+++ packages/orq-rc/src/orq_ai_sdk/models/getallmemorystoresop.py
@@ -89,7 +89,7 @@ class GetAllMemoryStoresDataTypedDict(TypedDict):
     r"""The user ID of the creator"""
     updated_by_id: NotRequired[Nullable[str]]
     r"""The user ID of the last updater"""
-    ttl: NotRequired[float]
+    ttl: NotRequired[Nullable[float]]
     r"""The default time to live of every memory document created within the memory store. Useful to control if the documents in the memory should be store for short or long term."""
 
 
@@ -117,13 +117,13 @@ class GetAllMemoryStoresData(BaseModel):
     updated_by_id: OptionalNullable[str] = UNSET
     r"""The user ID of the last updater"""
 
-    ttl: Optional[float] = None
+    ttl: OptionalNullable[float] = UNSET
     r"""The default time to live of every memory document created within the memory store. Useful to control if the documents in the memory should be store for short or long term."""
 
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = set(["created_by_id", "updated_by_id", "ttl"])
-        nullable_fields = set(["created_by_id", "updated_by_id"])
+        nullable_fields = set(["created_by_id", "updated_by_id", "ttl"])
         serialized = handler(self)
         m = {}