+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1960,7 +1960,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KTV1AZWBNE075G2PTT3XFS3R"
+    id: Optional[str] = "01KTVC27KQNF3061G313WWDKRX"
 
     description: Optional[str] = None