+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTV1B07PP634TWDJJM6P5888"
+        "01KTVC27ZM70Q8HZ819Z8SJYAA"
     )
     r"""The unique identifier of the data source"""